Add intermediary component to map the columns of the imported CSV.
This aims at solving the problem of unintuitivly resetting the form every time the column mapping changes because those changes are now only possible in another form/view.
